Move setting helpers into new helpers.php file; Refactoring to put settings into the app config
