+++ ProcessMaker/Observers/SettingObserver.php
@@ -70,9 +70,13 @@ class SettingObserver
     public function saved(Setting $setting)
     {
         try {
-            cache_settings(true);
+            refresh_artisan_caches(clear_artisan_caches());
         } catch (Exception $exception) {
-            Log::error('Could not cache configuration.', [$exception]);
+            Log::error('Could not cache configuration.', [
+                'message' => $exception->getMessage(),
+                'file' =>$exception->getFile(),
+                'line' => $exception->getLine(),
+            ]);
         }
     }
 
@@ -85,9 +89,13 @@ class SettingObserver
     public function deleted(Setting $setting)
     {
         try {
-            cache_settings(true);
+            refresh_artisan_caches(clear_artisan_caches());
         } catch (Exception $exception) {
-            Log::error('Could not cache configuration.', [$exception]);
+            Log::error('Could not cache configuration.', [
+                'message' => $exception->getMessage(),
+                'file' =>$exception->getFile(),
+                'line' => $exception->getLine(),
+            ]);
         }
     }
 }
