+++ ProcessMaker/Providers/SettingServiceProvider.php
@@ -2,24 +2,43 @@
 
 namespace ProcessMaker\Providers;
 
+use ProcessMaker\Models\Setting;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Contracts\Config\Repository as RepositoryContract;
 
 class SettingServiceProvider extends ServiceProvider
 {
-    public function register()
-    {
-        require_once(app_path('/Helpers/SettingsHelper.php'));
-    }
-
     /**
-     * Bootstrap services.
+     * Bootstrap settings into the global app configuration
      *
-     * @throws \Exception
+     * @return void
+     *
+     * @throws \Psr\Container\ContainerExceptionInterface
+     * @throws \Psr\Container\NotFoundExceptionInterface
      */
     public function boot()
     {
-        if (!config('settings_loaded')) {
-            cache_settings();
+        if (!$this->app->configurationIsCached()) {
+            $this->loadSettingsFromDatabase($this->app->get('config'));
+        }
+    }
+
+    /**
+     * Bind setting keys and config values to the global app configuration
+     *
+     * @param  \Illuminate\Contracts\Config\Repository  $repository
+     *
+     * @return void
+     */
+    protected function loadSettingsFromDatabase(RepositoryContract $repository): void
+    {
+        if (!Schema::hasTable('settings')) {
+            return;
+        }
+
+        foreach (Setting::select('id', 'key', 'config')->get() as $setting) {
+            $repository->set($setting->key, $setting->config);
         }
     }
 }
